requirements.txt
absl-py==2.3.1
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
brotlicffi
certifi
cffi
charset-normalizer
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
eval_type_backport==0.2.2
faster-coco-eval==1.7.0
filelock
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2
grpcio==1.76.0
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
Jinja2
joblib==1.5.2
kiwisolver==1.4.9
lightly==1.5.22
lightly-utils==0.0.2
lightly_train==0.11.4
lightning-utilities==0.15.2
Markdown==3.9
MarkupSafe
matplotlib==3.10.7
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
multidict==6.7.0
networkx
numpy==2.0.1
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow
platformdirs==4.5.0
polars==1.35.1
polars-runtime-32==1.35.1
propcache==0.4.1
protobuf==3.20.3
psutil==7.1.2
pyarrow==22.0.0
pycocotools==2.0.10
pycparser
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
PySocks
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.5
PyYAML
regex==2025.10.23
requests
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.43.0
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
stringzilla==4.2.3
streamlit              # your app
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.21
tokenizers==0.22.1
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.8.2
torchvision==0.20.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions
ultralytics==8.3.223
ultralytics-thop==2.0.18
urllib3
wandb==0.22.3
Werkzeug==3.1.3
win_inet_pton
yarl==1.22.0

# Extra packages used earlier
gradio
opencv-python
Pillow
numpy
